requirements.txt
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
pytest>=7.0.0
pytest-cov>=4.0.0
flake8>=6.0.0
black>=23.0.0
numpy>=1.21.0
setuptools>=65.0.0
build>=0.10.0
